refactor: hooks to be called via aicommits
